definitions/raster-calculations/containers/reproject-raster/requirements.txt
numpy
python-decouple
rasterio